Much faster simulation.
Simulation.run() now has only one main for. But it wasn't the main
performance problem. It was Particle.interact(...), which made it's own
Random and that's why it was the slowest method after run(). Now, it
uses "main" Random and program run much, much faster - 60 iterations
take maybe 15-20s, when previous record was 2min!